arkavo-llm: split spec generation into submodule + seed spec cache from prior tokens
Code review followups to 20a3e1f4:

- llamacpp_streaming.rs hit 1.7k lines; extract spec-only functions
  (generate_tokens_with_spec, generate_tokens_pooled_with_spec,
  emit_token) into llamacpp_streaming/spec.rs. Parent module keeps
  the dispatchers and baseline paths.

- ContextReuseOptions gains prior_tokens: Vec<i32>. When resuming via
  start_position, the spec ngram cache now scans the prior turn's
  output tokens in addition to the new prompt segment, restoring the
  Infer2 multi-turn optimization that was silently degraded. Callers
  that don't yet thread prior_tokens pass Vec::new() with a comment.
